+++ Store-management-system/foco-store-server/src/main/java/com/bookstore/app/services/UserService.java
@@ -4,8 +4,11 @@ import java.util.List;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.bookstore.app.ExceptionHandling.TextBooksNotFoundException;
 import com.bookstore.app.models.UserInfo;
 import com.bookstore.app.repositories.UserRepo;
 
@@ -19,23 +22,44 @@ public class UserService {
 		return userRepo.findAll();		
 	}
 
-	public UserInfo getUser(UUID userId){ 
-		return userRepo.findById(userId).get();
+	public ResponseEntity<UserInfo> getUser(UUID id){ 
+		if(!availabilityCheck(id)) {
+			throw new TextBooksNotFoundException();
+		}
+		else {
+			return new ResponseEntity<>(userRepo.findById(id).get(), HttpStatus.OK);
+		}
 	}
 
-	public void addUser(UserInfo userInfo){	
+	public ResponseEntity<UserInfo> addUser(UserInfo userInfo){	
 		userRepo.save(userInfo);
+		return new ResponseEntity<>(userInfo, HttpStatus.OK);
 	}
 
-	public void updateUser(UserInfo userInfo, UUID id) {	
-		UserInfo userToupdate = userRepo.findById(id).get();
-		userToupdate = userInfo;
-		userRepo.save(userToupdate);
-		return;
+	public ResponseEntity<UserInfo> updateUser(UserInfo userInfo, UUID id) {	
+		if(!availabilityCheck(id)) {
+			throw new TextBooksNotFoundException();
+		}
+		else {
+			UserInfo userToUpdate = userRepo.findById(id).get();
+			userToUpdate = userInfo;
+			userRepo.save(userToUpdate);
+		    return new ResponseEntity<>(userRepo.findById(id).get(), HttpStatus.OK);
+		}	
 	}
 
-	public void deleteUser(UUID id){	
-		userRepo.deleteById(id);
+	public ResponseEntity<String> deleteUser(UUID id){	
+		if(!availabilityCheck(id)) {
+			throw new TextBooksNotFoundException();
+		}
+		else {
+			userRepo.deleteById(id);
+			return new ResponseEntity<>("Book with id " + id + " deleted", HttpStatus.OK);
+		}
+	}
+	
+	public boolean availabilityCheck(UUID id) {
+		return userRepo.existsById(id);
 	}
 
 }
